Split process files into create, update, delete
